Add death event
