+++ server/event/handler.h
@@ -12,5 +12,6 @@
  * @return false
  */
 bool meteor_handler(server_t *srv, const event_t *event);
+bool player_death_handler(server_t *srv, const event_t *event);
 
 #endif /* !HANDLER_H_ */
